Extract out DownloadManager from download service.
This allows injecting a different download manager by the caller
and adds more flexibility when using anytime as a dependency.
