+++ lib/ui/anytime_podcast_app.dart
@@ -16,6 +16,7 @@ import 'package:anytime/repository/repository.dart';
 import 'package:anytime/repository/sembast/sembast_repository.dart';
 import 'package:anytime/services/audio/audio_player_service.dart';
 import 'package:anytime/services/audio/mobile_audio_service.dart';
+import 'package:anytime/services/download/download_manager.dart';
 import 'package:anytime/services/download/download_service.dart';
 import 'package:anytime/services/download/mobile_download_service.dart';
 import 'package:anytime/services/podcast/mobile_podcast_service.dart';
@@ -59,7 +60,7 @@ class AnytimePodcastApp extends StatefulWidget {
   AnytimePodcastApp(this.mobileSettingsService)
       : repository = SembastRepository(),
         podcastApi = MobilePodcastApi() {
-    downloadService = MobileDownloadService(repository: repository);
+    downloadService = MobileDownloadService(repository: repository, downloadManager: FlutterDownloaderManager());
     podcastService = MobilePodcastService(api: podcastApi, repository: repository, settingsService: mobileSettingsService);
     audioPlayerService = MobileAudioPlayerService(repository: repository);
     settingsBloc = SettingsBloc(mobileSettingsService);
